pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beeperclaw"
version = "0.1.0"
description = "AI coding agent accessible from anywhere via Beeper/Matrix"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
keywords = [
    "ai",
    "coding",
    "agent",
    "matrix",
    "beeper",
    "opencode",
    "automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "simplematrixbotlib>=2.10.0",
    "matrix-nio[e2e]>=0.24.0",
    "aiohttp>=3.9.0",
    "markdown>=3.6",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",
    "pre-commit>=3.6.0",
]
github = [
    "PyGithub>=2.1.0",
]

[project.scripts]
beeperclaw = "beeperclaw.cli:main"

[project.urls]
Homepage = "https://github.com/your-username/beeperclaw"
Documentation = "https://github.com/your-username/beeperclaw#readme"
Repository = "https://github.com/your-username/beeperclaw"
Issues = "https://github.com/your-username/beeperclaw/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/beeperclaw"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["markdown", "nio", "nio.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
